BL-2565 Cleanup after code review
